Refactor StreamEncryption and fix Interop types

+++ src/FsSodium/Random.fs
@@ -2,5 +2,5 @@ module FsSodium.Random
 
 let bytes count =
     let result = Array.zeroCreate count
-    Interop.randombytes_buf (result, uint64 count)
+    Interop.randombytes_buf (result, uint32 count)
     result
